finished refactor and fixed broken rules
